# docs/api/openapi.yml
openapi: 3.0.3
info:
  title: ADHD Productivity System API
  description: |
    A comprehensive API for managing tasks, focus sessions, and productivity metrics specifically designed for individuals with ADHD.
    
    ## Key Features
    - **ADHD-Centered Design**: All endpoints are optimized for users with ADHD
    - **Focus Session Management**: Pomodoro-style focus sessions with ADHD-friendly customizations
    - **Task Decomposition**: Smart task breaking for executive function support
    - **Real-time Updates**: WebSocket support for live session synchronization
    - **Analytics & Insights**: Productivity tracking with ADHD-specific metrics
    
    ## Authentication
    This API uses JWT (JSON Web Tokens) for authentication. Include the token in the Authorization header:
    `Authorization: Bearer <your-jwt-token>`
    
    ## Rate Limiting
    - Standard endpoints: 100 requests/minute per user
    - Authentication endpoints: 5 requests/minute per IP
    - File uploads: 10 requests/minute per user
    
    ## ADHD-Specific Considerations
    - All response times are optimized for attention management
    - Error messages are clear and actionable
    - Minimal cognitive load in API design
    - Support for executive function assistance
  version: 1.0.0
  contact:
    name: ADHD Productivity System Team
    email: api-support@adhd-productivity.dev
    url: https://docs.adhd-productivity.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
...
security:
  - bearerAuth: []
paths:
...
  /tasks:
    get:
      tags:
        - Tasks
      summary: Get user tasks
      description: |
        Retrieve tasks with ADHD-friendly filtering and sorting options.
        
        **ADHD Features:**
        - Priority-based sorting for executive function support
        - Energy-level filtering for task matching
        - Estimated duration for time-blocking
        - Context switching indicators
      operationId: getTasks
      parameters:
        - name: status
          in: query
          description: Filter by task status
          schema:
            type: string
            enum: [todo, in_progress, completed, cancelled]
        - name: priority
          in: query
          description: Filter by priority level
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: energyLevel
          in: query
          description: Filter by required energy level
          schema:
            type: string
            enum: [low, medium, high]
        - name: estimatedDuration
          in: query
          description: Filter by estimated duration (minutes)
          schema:
            type: integer
            minimum: 1
            maximum: 480
        - name: dueDate
          in: query
          description: Filter by due date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: tags
          in: query
          description: Filter by tags (comma-separated)
          schema:
            type: string
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Tasks
      summary: Create a new task
      description: |
        Create a new task with ADHD-specific features:
        - Automatic task decomposition suggestions
        - Energy level assessment
        - Context switching warnings
        - Dopamine reward integration
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
            examples:
              simple_task:
                summary: Simple task creation
                value:
                  title: "Write project proposal"
                  description: "Draft the Q1 project proposal for management review"
                  priority: "high"
                  estimatedDuration: 60
                  dueDate: "2024-12-25T10:00:00Z"
              adhd_optimized:
                summary: ADHD-optimized task
                value:
                  title: "Research market trends"
                  description: "Gather data on current market trends for the proposal"
                  priority: "medium"
                  estimatedDuration: 30
                  energyLevel: "medium"
                  tags: ["research", "quick-win"]
                  contextSwitchWarning: true
                  rewardType: "completion_badge"
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid task data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tasks/{taskId}:
    get:
      tags:
        - Tasks
      summary: Get specific task
      description: Retrieve detailed information about a specific task
      operationId: getTask
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task unique identifier
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Tasks
      summary: Update task
      description: Update an existing task with validation for ADHD-specific features
      operationId: updateTask
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task unique identifier
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid update data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Tasks
      summary: Delete task
      description: Delete a task (with confirmation for ADHD users to prevent accidental deletion)
      operationId: deleteTask
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task unique identifier
          schema:
            type: string
            format: uuid
        - name: confirm
          in: query
          description: Confirmation flag (required for ADHD safety)
          required: true
          schema:
            type: boolean
      responses:
        '204':
          description: Task deleted successfully
        '400':
          description: Missing confirmation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Focus Session Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title
          example: "Write project proposal"
        description:
          type: string
          maxLength: 2000
          description: Detailed task description
          example: "Draft the Q1 project proposal for management review"
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
          description: Task priority level
          example: "high"
        estimatedDuration:
          type: integer
          minimum: 1
          maximum: 480
          description: Estimated duration in minutes
          example: 60
        energyLevel:
          type: string
          enum: [low, medium, high]
          description: Required energy level
          example: "medium"
        dueDate:
          type: string
          format: date-time
          description: Task due date
          example: "2024-12-25T10:00:00Z"
        tags:
          type: array
          items:
            type: string
          description: Task tags for organization
          example: ["work", "proposal", "urgent"]
        contextSwitchWarning:
          type: boolean
          default: false
          description: Show context switch warning
          example: true
        rewardType:
          type: string
          enum: [none, points, badge, celebration]
          description: Dopamine reward type
          example: "badge"
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [todo, in_progress, completed, cancelled]
        estimatedDuration:
          type: integer
          minimum: 1
          maximum: 480
        actualDuration:
          type: integer
          minimum: 0
        energyLevel:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
          maxLength: 1000
          description: Additional notes or comments
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440001"
        title:
          type: string
          example: "Write project proposal"
        description:
          type: string
          example: "Draft the Q1 project proposal for management review"
        priority:
          type: string
          enum: [low, medium, high, urgent]
          example: "high"
        status:
          type: string
          enum: [todo, in_progress, completed, cancelled]
          example: "todo"
        estimatedDuration:
          type: integer
          example: 60
        actualDuration:
          type: integer
          nullable: true
          example: null
        energyLevel:
          type: string
          enum: [low, medium, high]
          example: "medium"
        dueDate:
          type: string
          format: date-time
          nullable: true
          example: "2024-12-25T10:00:00Z"
        tags:
          type: array
          items:
            type: string
          example: ["work", "proposal", "urgent"]
        createdAt:
          type: string
          format: date-time
          example: "2024-12-22T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-12-22T10:00:00Z"
        completedAt:
          type: string
          format: date-time
          nullable: true
          example: null
        focusSessionsCount:
          type: integer
          description: Number of focus sessions completed for this task
          example: 3
        rewardEarned:
          type: boolean
          description: Whether dopamine reward was earned
          example: false
    TaskResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Task'
        message:
          type: string
          example: "Task retrieved successfully"
        timestamp:
          type: string
          format: date-time
          example: "2024-12-22T10:00:00Z"
    TaskListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            tasks:
              type: array
              items:
                $ref: '#/components/schemas/Task'
            pagination:
              type: object
              properties:
                page:
                  type: integer
                  example: 1
                limit:
                  type: integer
                  example: 20
                total:
                  type: integer
                  example: 45
                totalPages:
                  type: integer
                  example: 3
        message:
          type: string
          example: "Tasks retrieved successfully"
        timestamp:
          type: string
          format: date-time
          example: "2024-12-22T10:00:00Z"

    # Focus Session Schemas
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        data:
          type: object
          nullable: true
          example: null
        message:
          type: string
          description: Human-readable error message
          example: "Invalid input data"
        error:
          type: string
          description: Technical error details
          example: "Validation failed for field 'email'"
        timestamp:
          type: string
          format: date-time
          example: "2024-12-22T10:00:00Z"
        validationErrors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "email"
              message:
                type: string
                example: "Must be a valid email address"
              code:
                type: string
                example: "INVALID_FORMAT"
        errorCode:
          type: string
          description: Application-specific error code
          example: "VALIDATION_ERROR"
        requestId:
          type: string
          description: Request tracking ID for support
          example: "req_123456789"
...